Allow Edit button to stay active after submit
After submitting, the edit button was responding to the refresh, making subsequent edits frustrating since it would automatically switch to read-only. This adjusts that logic.

However, the side effect is that the resourceData may not have yet caught up with the request, so clicking Submit again immediately after editing for a second time will result in a parsing error, but it seems the chances someone will click on the "Submit" button so quickly may be relatively low.