Allow polling for all but violations tab
Clicking the violations tab would stop polling for other tabs since the refresh state was being passed now that refresh controls were present. The violations tab has its own query interval based on the refresh, which is why polling is stopped for auto refresh but the controls are still relevant.