Fix filter functionality
Locally the filters were not displaying because our filters are saved as sets and sets cannot be stringified for the sessionStorage object. The fix converts these sets to arrays so that they can be stringified for the sessionStorage. Additionally, we're deleting the value before setting it when it appears that just setting it is sufficient.