Clone prop.rows
prop.rows was being assigned directly to sortRows. The problem with this is that if columns were being sorted by rows with duplicate data, a refresh would cause the rows to re-sort to their original order. This goes ahead and sorts it that way to begin with so that it doesn't unexpectedly re-sort on a prop refresh.

Additionally, both Array.filter() and Array.sort() modify arrays in place, so we can assign it up front.