Remove tab from overview page
With this change, we have logic to not show breadcrumbs/tabs/buttons, so the other `else` statement to have a simple header was both unused and unnecessary.

Additionally, narrowing the secondary header was considered but other pages that don't have tabs (like the cluster drilldown) also have a wide header, so I made the decision for simplicity and consistency to keep a secondary header of uniform height.