Fix filter issue on card click
When a card is clicked, the toggle is set to disappear and the session cookie was set using `grcCardChoice`. However, `grcCardChoice` was recently updated to use indexes, so this adjusts the URL to update without the toggle filter applied if it exists and also updates to use the label for the card instead of the index.