Fix checkbox behavior
When a specification template was selected, we added its name to a wasSet array inside of the form element that it updated. However, we were not adding the template name to wasSet if the replacement wasn't made. That means that when a template was unselected, template not added would add their name to wasSet and replace the value. This fix adds the template name to wasSet regardless so that any customization does not get unexpectedly overridden.

Additionally, we did not have logic to remove template names from wasSet if they were unselected except when all template names were removed. This adds this logic so that if a template is unselected and then reselected, its replacements will be made once again.