Fix YAML crashing on custom spec update - 7380 (#316)
* Update Specs path; add policy-template to template

* Update template behaviors to fix crashing editor

The editor was crashing because custom specifications were not being parsed correctly because the YAML path was incorrect, returning strings instead of an expected array. This corrects the path so that the `active` and `userData` keys in `controlData` are updated properly.

* Update behavior so that placeholder is correct

When a custom template was deleted, we were setting `active` to be the full custom template. However, this also meant that the placeholder was being updated to also be the entire template. By setting an empty array for `active`, the placeholder is restored.

* Fix interactions between YAML editor and UI

- Defaults name to an empty string if the field is erased
- Erase Specifications menu item if specification templates are erased
  - Additionally, handle empty template list to revert to empty array
- Return custom 'YAML returned null' error instead of mysterious '$synced is null' with empty YAML
- Now that templates are separate from remediation/disable spec, remove that unused code
- Update template so that `policy-templates:` always appears

* Grab the parent key from `reverse`

It was coded as a static string, so this could be more maintainable if something changes

* Update E2E tests

Co-authored-by: Alex Luo <55596988+ChunxiAlexLuo@users.noreply.github.com>
Co-authored-by: Yu Cao <ycao@redhat.com>