Simplify RegEx checks for performance
The multiple nested groups in the Kubernetes RFC 1123 RegEx was causing a browser stack overflow (specifically, "Too much recursion" in Firefox), so this change simplifies the check.

Additionally, it also adds the check for the editor so that the editor is decorated with the error.

Finally, it updates the state on render with the full RegEx so that if "Create" is clicked, an error appears instead of attempting to create it. (React doesn't like this, but I'm not sure how to refactor this offhand.)