Remove the unique name check
We were checking for a unique name and updating the name if it wasn't unique. Since the create page now updates policies with the same name and we have a notification when the name is not unique (and this also resulted in an edge case bug), we'll just remove it.