introduce way of loosening checks when testsuite not run standalone
There are tests that expect no policy being present on a cluster
before test starts in order to be able to do precide content checks.
When testsuite is run in parallel with other tests, this condition
is not met.
This commit introduces loosening of such checks which is enabled
by setting the environment variable
STANDALONE_TESTSUITE_EXECUTION
(resp. CYPRESS_STANDALONE_TESTSUITE_EXECUTION)
to 'FALSE'.