Implement strict search
If a user has something like cluster1 and cluster12, the clusterFilter should be able to show them only the cluster they clicked on rather than all clusters that have the substring given. This means that if the search filter changes, we should wipe out the query from the URL and let the user search at will in the table.

Signed-off-by: Dale Haiducek <dhaiduce@redhat.com>