Use AcmTable in byTemplate tables in status view
Slowly removing the uses of the GRC implementation of PatternFlyTable.

As a result of how the searchbar-controlling function is implemented,
the search query from the "Clusters" tab is carried over into every
table under the "Templates" tab.

The tab-switcher between the Clusters and Templates view is repeated on
each table under the "Templates" tab, which might be wrong, but moving
it outside of the table would make it different than how it is on the
"Clusters" tab, which seems to be the 'correct' way to do it...

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12036

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>