Handle more resize events
The window `resize` event was not firing when the left nav was shown or
hidden, so the YamlEditor on this page was not updating. Since the yaml
is now in a card, the effect was more obvious than it was previously.
The react-resize-observer uses something non-obvious to detect the
changes, so I couldn't just add a new listener. In fact, the onResize
event is not even sufficient in all cases here, but by using onReflow,
even the cases we miss will get fixed if/when the user scrolls.

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>