Use sortable: true instead of transform
AcmTable will automatically add the sortable transform to the table
items if we give it a sort property. Using a sort property will let it
sort on things that aren't exactly what is in the cell, which is
necessary if the cell has a link or something in it.

If both AcmTable and our TableDefs add the sortable transform, then the
table header will have two sortable arrows, which is not good.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12761

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>