Use browser ResizeObserver API for yaml resizing
Previously a React component was used to detect size changes to the yaml
editor in the details view page, but there were still some corner cases
where this did not correctly detect the change and update the layout.
It turns out (some) browsers have an API that works better, which we
can use instead.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12614

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>