Prompt user to install Ansible Operator in modal
Ansible automations require the ansible operator to be installed in, but
we currently don't check that before allowing the UI to try to create a
Job. This will create an Alert at the top of the ansible automation
modal if the operator is not installed, linking the user to the
OperatorHub where they could install it. It also removes the "Save" and
"Cancel" buttons if the oeprator is not installed.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/12826

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>