Skip RBAC tests on ROSA clusters (#701)
* Expose rbac-setup as an npm script

This could make it easier to re-setup the rbac, if it was removed from
the cluster, or if it had an error. For example, cypress would now be
able to set up the IDP with `cy.exec('npm run rbac:setup')`.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Exclude rbac tests on ROSA clusters

Per Slack conversation with cicd, it may be easiest to skip RBAC tests
on ROSA clusters for now, since the `htpasswd` IDP that we usually use
will be periodically removed on those clusters.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/14250

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Fix idp setup test

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Standardize quotes around bash variables

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>