Remove the Ansible automation modal test
This has been flaky for years and the Console squad and GRC squad have
not figured out how to make this work consistently.

A new approach was taken in the 2.5 console rewrite, so this test will
be removed in 2.4. The QE E2E tests handle this test case, so there is
still coverage before a new release.

Relates:
https://github.com/stolostron/backlog/issues/17369
https://github.com/stolostron/backlog/issues/24097

Signed-off-by: mprahl <mprahl@users.noreply.github.com>